- Add CampaignDB to open an ACR recording database at first event, to record files one by one as they appear, and close at the end. - Rename table 'bpfiles' to 'bpdataset' to be similar to ACA db. - Add table 'info' similar to ACA. This one is named ACR, ADIOS Campaign Record. - Change ctime field to REAL although this does not seem to make a difference in sqlite3. - Change ACA and ACR version to 1.1

+++ source/adios2/engine/campaign/CampaignData.h
@@ -41,7 +41,7 @@ struct CampaignBPFile
     bool compressed;
     size_t lengthOriginal;
     size_t lengthCompressed;
-    long ctime;
+    double ctime;
 };
 
 struct CampaignBPDataset
